Updated project
Added the correctly flow of the application using fragments.
Added a new theme framework to simplify the design work.
Made welcome pages just left add the string constrains to the resource file and sync it with the textviews.
Made the login page, left add register click handler.
Next thing I'll do is register activity and home page to work as good as we want.